Fix and reorganize integration tests
